Update / Refactoring
- changed text styles
- reorganized ui packages
- ...
