+++ XYZReader/src/main/java/com/example/xyzreader/ui/custom/DynamicHeightNetworkImageView.java
@@ -1,4 +1,4 @@
-package com.example.xyzreader.ui;
+package com.example.xyzreader.ui.custom;
 
 import android.content.Context;
 import android.util.AttributeSet;
