+++ XYZReader/src/main/java/com/example/xyzreader/ui/custom/ObservableScrollView.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.example.xyzreader.ui;
+package com.example.xyzreader.ui.custom;
 
 import android.content.Context;
 import android.os.Parcelable;
